Completed refactor of HeatMap > EnergyMap
