+++ energy-app-react/src/energymap/EnergyMapView.js
