+++ energy-app-react/src/energymap/Loader.js
